+++ crates/criticalup-cli/src/arg.rs
@@ -35,20 +35,30 @@ impl<'a> Instrumentation {
     }
 
     pub(crate) async fn setup<'b: 'a>(&'b self, binary_name: &str) -> Result<(), crate::Error> {
-        let fmt_layer = self.fmt_layer();
         let filter_layer = self.filter_layer(binary_name)?;
 
-        tracing_subscriber::registry()
-            .with(filter_layer)
-            .with(fmt_layer)
-            .try_init()?;
+        if self.verbose != 0 {
+            let fmt_layer = self.verbose_fmt_layer();
+            tracing_subscriber::registry()
+                .with(filter_layer)
+                .with(fmt_layer)
+                .try_init()?;
+        } else {
+            let fmt_layer = self.fmt_layer();
+            tracing_subscriber::registry()
+                .with(filter_layer)
+                .with(fmt_layer)
+                .try_init()?;
+        }
 
         tracing::trace!("Instrumentation initialized");
 
         Ok(())
     }
 
-    pub(crate) fn fmt_layer<S>(&self) -> impl tracing_subscriber::layer::Layer<S>
+    /// Set up a 'pretty' formatter that displays structure span information, timestamps,
+    /// line numbers/files, etc.
+    pub(crate) fn verbose_fmt_layer<S>(&self) -> impl tracing_subscriber::layer::Layer<S>
     where
         S: tracing::Subscriber + for<'span> tracing_subscriber::registry::LookupSpan<'span>,
     {
@@ -56,6 +66,20 @@ impl<'a> Instrumentation {
             .with_ansi(std::io::stderr().is_terminal())
             .with_writer(std::io::stderr)
             .pretty()
+    }
+
+    /// Set up a basic, simple logger that doesn't emit more than it needs.
+    pub(crate) fn fmt_layer<S>(&self) -> impl tracing_subscriber::layer::Layer<S>
+    where
+        S: tracing::Subscriber + for<'span> tracing_subscriber::registry::LookupSpan<'span>,
+    {
+        tracing_subscriber::fmt::Layer::new()
+            .with_ansi(std::io::stderr().is_terminal())
+            .with_writer(std::io::stderr)
+            .without_time()
+            .with_file(self.verbose != 0)
+            .with_line_number(self.verbose != 0)
+            .with_target(self.verbose != 0)
     }
 
     pub(crate) fn filter_layer(&self, binary_name: &str) -> Result<EnvFilter, crate::Error> {